Simplify db getter and extractor hierarchy
