+++ instrumentation-api-incubator/src/main/java/io/opentelemetry/instrumentation/api/incubator/semconv/db/SqlClientAttributesGetter.java
@@ -18,8 +18,7 @@ import javax.annotation.Nullable;
  * from the attribute methods, but implement as many as possible for best compliance with the
  * OpenTelemetry specification.
  */
-public interface SqlClientAttributesGetter<REQUEST>
-    extends DbClientCommonAttributesGetter<REQUEST> {
+public interface SqlClientAttributesGetter<REQUEST> extends DbClientAttributesGetter<REQUEST> {
 
   /**
    * Get the raw SQL statement. The value returned by this method is later sanitized by the {@link
